test: move getConnections test to its own file
node-net-server.test.ts has several pre-existing tests with 100ms
setTimeout races (setTimeout(closeAndFail, 100)) that flake under
ASAN load. Moving the getConnections drain test to its own file
keeps it isolated from those, matching the per-topic file layout
already used in test/js/node/net/.
